实现Comment后台功能

+++ src/main/java/com/tanzhou/controller/PublishController.java
@@ -20,7 +20,7 @@ public class PublishController {
     @Autowired
     private QuestionService questionService;
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name="id")Integer id,Model model){
+    public String edit(@PathVariable(name="id")Long id,Model model){
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title",question.getTitle());
         model.addAttribute("description",question.getDescription());
@@ -38,7 +38,7 @@ public class PublishController {
     public String doPublish(@RequestParam(value = "title",required = false)String title,
                             @RequestParam(value = "description",required = false)String description,
                             @RequestParam(value = "tag",required = false)String tag,
-                            @RequestParam("id") Integer id, HttpServletRequest request,
+                            @RequestParam("id") Long id, HttpServletRequest request,
                             Model model){
         model.addAttribute("title",title);
         model.addAttribute("description",description);
